글쓰기에서 board.save와  fileup.save 구현, boardfile 구현중

+++ src/main/java/shoes/skream/project/repository/won/CategoryRepositoryWon.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import shoes.skream.project.domain.Category;
 
-public interface CategoryRepositoryWon extends JpaRepository<Category, Long>{
+public interface CategoryRepositoryWon extends JpaRepository<Category, Integer>{
     
 }
